# docs/swagger.yaml
definitions:
...
  handlers.CreateNoteRequest:
    properties:
      content:
        example: This is the content of my first note
        minLength: 1
        type: string
      title:
        example: My First Note
        maxLength: 255
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
...
  handlers.NoteResponse:
    properties:
      content:
        example: This is the content of my first note
        type: string
      created_at:
        example: 1672531200
        type: integer
      id:
        example: 1
        type: integer
      title:
        example: My First Note
        type: string
      updated_at:
        example: 1672531200
        type: integer
    type: object
  handlers.NotesListResponse:
    properties:
      count:
        example: 5
        type: integer
      limit:
        example: 10
        type: integer
      notes:
        items:
          $ref: '#/definitions/handlers.NoteResponse'
        type: array
      page:
        example: 1
        type: integer
      total:
        example: 25
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
...
  handlers.UpdateNoteRequest:
    properties:
      content:
        example: This is the updated content of my note
        minLength: 1
        type: string
      title:
        example: Updated Note Title
        maxLength: 255
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
...
info:
  contact: {}
paths:
...
  /notes:
    get:
      consumes:
      - application/json
      description: Get all notes for the authenticated user with pagination and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of notes per page
        in: query
        name: limit
        type: integer
      - description: Search term for notes title or content
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notes retrieved successfully
          schema:
            $ref: '#/definitions/handlers.NotesListResponse'
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all notes
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Create a new note for the authenticated user
      parameters:
      - description: Note creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Note created successfully
          schema:
            $ref: '#/definitions/handlers.NoteResponse'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - Notes
  /notes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific note by ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Note deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid note ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Note not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a specific note
      tags:
      - Notes
    get:
      consumes:
      - application/json
      description: Get a specific note by ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Note retrieved successfully
          schema:
            $ref: '#/definitions/handlers.NoteResponse'
        "400":
          description: Invalid note ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Note not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a specific note
      tags:
      - Notes
    put:
      consumes:
      - application/json
      description: Update a specific note by ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      - description: Note update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Note updated successfully
          schema:
            $ref: '#/definitions/handlers.NoteResponse'
        "400":
          description: Invalid request data or note ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Note not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a specific note
      tags:
      - Notes
swagger: "2.0"